mdbook_driver

Module book

A tree structure representing a book.

Structs

Book

A tree structure representing a book.

BookItems

A depth-first iterator over the items in a book.

Chapter

The representation of a “chapter”, usually mapping to a single file on disk however it may contain multiple sub-chapters.

SectionNumber

A section number like “1.2.3”, basically just a newtype’d Vec<u32> with a pretty Display impl.

Enums

BookItem

Enum representing any type of item which can be added to a book.